I've made the --password argument optional to work better with the Rails template.
Here's the deal. I tried running the Rails template and it b0rked when it got to making the database because the password on mine is blank right now. At first, I thought it was a bug in production data that it wasn't prepared to handle a blank password. Error looked like:

gimme root sql password for db_setup:
script/db_setup: option `--password' requires an argument
/Users/gordon/.rvm/rubies/ruby-1.8.7-p302/lib/ruby/1.8/getoptlong.rb:403:in `set_error': option `--password' requires an argument (GetoptLong::MissingArgument)
									    from /Users/gordon/.rvm/rubies/ruby-1.8.7-p302/lib/ruby/1.8/getoptlong.rb:523:in `get_option'
									    from /Users/gordon/.rvm/rubies/ruby-1.8.7-p302/lib/ruby/1.8/getoptlong.rb:611:in `each'
									    from /Users/gordon/.rvm/rubies/ruby-1.8.7-p302/lib/ruby/1.8/getoptlong.rb:610:in `loop'
									    from /Users/gordon/.rvm/rubies/ruby-1.8.7-p302/lib/ruby/1.8/getoptlong.rb:610:in `each'
									    from script/db_setup:31

The actual deal is that the template calls the db_setup script with '--password'. Of course, '--password' is marked as 'REQUIRED_ARGUMENT' in the db_setup script, so if you pass a blank password you get the error above. If you don't pass '--password', then the template won't work because the db_setup script will prompt for a password in the background and cause the whole thing to hang. Consequently, I marked '--password' as 'OPTIONAL_ARGUMENT'. It's not ideal in the case that someone tries to use --password without an argument and doesn't understand why it doesn't work, but it does make the current situation work.

Other solutions are welcome.